Moved DFA-generating code to StateManager so that it can be accessed from anywhere in the program
